+++ packages/platform/test/Http/UrlParams.test.ts
@@ -4,6 +4,26 @@ import { Effect } from "effect"
 
 describe("UrlParams", () => {
   describe("makeUrl", () => {
+    it.effect("makes a URL", () =>
+      Effect.gen(function*(_) {
+        const url = yield* _(UrlParams.makeUrl("https://example.com/test", []))
+        assert.strictEqual(url.toString(), "https://example.com/test")
+      }))
+
+    it.effect("supports relative URLs", () =>
+      Effect.gen(function*(_) {
+        const originalLocation = globalThis.location
+
+        globalThis.location = {
+          origin: "https://example.com",
+          pathname: "/path/"
+        } as Location
+        const url = yield* _(UrlParams.makeUrl("test", []))
+        assert.strictEqual(url.toString(), "https://example.com/path/test")
+
+        globalThis.location = originalLocation
+      }))
+
     it.effect("does not throw if `location` is set to `undefined`", () =>
       Effect.gen(function*(_) {
         const originalLocation = globalThis.location
@@ -11,14 +31,39 @@ describe("UrlParams", () => {
         // `globalThis.location` is undefined
         // @ts-expect-error
         globalThis.location = undefined
-        let url = yield* _(UrlParams.makeUrl("http://example.com", []))
-        assert.strictEqual(url.toString(), "http://example.com/")
+        let url = yield* _(UrlParams.makeUrl("https://example.com", []))
+        assert.strictEqual(url.toString(), "https://example.com/")
 
         // `location` is not in globalThis
         // @ts-expect-error
         delete globalThis.location
         url = yield* _(UrlParams.makeUrl("http://example.com", []))
         assert.strictEqual(url.toString(), "http://example.com/")
+
+        globalThis.location = originalLocation
+      }))
+
+    it.effect("does not fail if `location` is partially defined", () =>
+      Effect.gen(function*(_) {
+        const originalLocation = globalThis.location
+
+        globalThis.location = { href: "" } as Location
+        const url1 = yield* _(UrlParams.makeUrl("https://example.com", []))
+        assert.strictEqual(url1.toString(), "https://example.com/")
+
+        globalThis.location = {
+          href: "",
+          origin: "https://example.com"
+        } as unknown as Location
+        const url2 = yield* _(UrlParams.makeUrl("https://example.com", []))
+        assert.strictEqual(url2.toString(), "https://example.com/")
+
+        globalThis.location = {
+          href: "",
+          pathname: "example_path"
+        } as unknown as Location
+        const url3 = yield* _(UrlParams.makeUrl("https://example.com", []))
+        assert.strictEqual(url3.toString(), "https://example.com/")
 
         globalThis.location = originalLocation
       }))